+++ scripts/lib/install/request.js
@@ -28,6 +28,7 @@ function parseInstallArgs(argv) {
     excludeComponentIds: [],
     languages: [],
     locale: null,
+    update: false,
   };
 
   for (let index = 0; index < args.length; index += 1) {
@@ -70,6 +71,8 @@ function parseInstallArgs(argv) {
       index += 1;
     } else if (arg === '--dry-run') {
       parsed.dryRun = true;
+    } else if (arg === '--update') {
+      parsed.update = true;
     } else if (arg === '--json') {
       parsed.json = true;
     } else if (arg === '--help' || arg === '-h') {
@@ -147,6 +150,7 @@ function normalizeInstallRequest(options = {}) {
     excludeComponentIds,
     legacyLanguages,
     configPath: config?.path || options.configPath || null,
+    update: Boolean(options.update),
   };
 }